Refactor DatafilesToolbar component and add TrashButton component
